+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/service/LoginHistoryMaintainServiceImpl.java
@@ -254,7 +254,8 @@ public class LoginHistoryMaintainServiceImpl implements LoginHistoryMaintainServ
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<LoginHistory> lookupAsList(String preset, Object[] objs, PagingInfo pagingInfo) throws ServiceException {
+    public List<LoginHistory> lookupAsList(String preset, Object[] objs, PagingInfo pagingInfo)
+            throws ServiceException {
         return presetLookupService.lookupAsList(preset, objs, pagingInfo);
     }