+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/service/ProtectorSupportMaintainServiceImpl.java
@@ -168,7 +168,8 @@ public class ProtectorSupportMaintainServiceImpl implements ProtectorSupportMain
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<StringIdKey> batchInsertIfNotExists(@SkipRecord List<ProtectorSupport> entities) throws ServiceException {
+    public List<StringIdKey> batchInsertIfNotExists(@SkipRecord List<ProtectorSupport> entities)
+            throws ServiceException {
         return crudService.batchInsertIfNotExists(entities);
     }
 
@@ -222,7 +223,8 @@ public class ProtectorSupportMaintainServiceImpl implements ProtectorSupportMain
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public PagedData<ProtectorSupport> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws ServiceException {
+    public PagedData<ProtectorSupport> lookup(String preset, Object[] objs, PagingInfo pagingInfo)
+            throws ServiceException {
         return presetLookupService.lookup(preset, objs, pagingInfo);
     }