+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/service/ProtectorVariableMaintainServiceImpl.java
@@ -129,7 +129,8 @@ public class ProtectorVariableMaintainServiceImpl implements ProtectorVariableMa
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<ProtectorVariableKey> batchInsert(@SkipRecord List<ProtectorVariable> elements) throws ServiceException {
+    public List<ProtectorVariableKey> batchInsert(@SkipRecord List<ProtectorVariable> elements)
+            throws ServiceException {
         return crudService.batchInsert(elements);
     }
 
@@ -151,7 +152,8 @@ public class ProtectorVariableMaintainServiceImpl implements ProtectorVariableMa
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<ProtectorVariable> batchGetIfExists(@SkipRecord List<ProtectorVariableKey> keys) throws ServiceException {
+    public List<ProtectorVariable> batchGetIfExists(@SkipRecord List<ProtectorVariableKey> keys)
+            throws ServiceException {
         return crudService.batchGetIfExists(keys);
     }
 
@@ -160,7 +162,8 @@ public class ProtectorVariableMaintainServiceImpl implements ProtectorVariableMa
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<ProtectorVariableKey> batchInsertIfExists(@SkipRecord List<ProtectorVariable> elements) throws ServiceException {
+    public List<ProtectorVariableKey> batchInsertIfExists(@SkipRecord List<ProtectorVariable> elements)
+            throws ServiceException {
         return crudService.batchInsertIfExists(elements);
     }
 
@@ -168,7 +171,8 @@ public class ProtectorVariableMaintainServiceImpl implements ProtectorVariableMa
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<ProtectorVariableKey> batchInsertIfNotExists(@SkipRecord List<ProtectorVariable> entities) throws ServiceException {
+    public List<ProtectorVariableKey> batchInsertIfNotExists(@SkipRecord List<ProtectorVariable> entities)
+            throws ServiceException {
         return crudService.batchInsertIfNotExists(entities);
     }
 
@@ -190,7 +194,8 @@ public class ProtectorVariableMaintainServiceImpl implements ProtectorVariableMa
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<ProtectorVariableKey> batchInsertOrUpdate(@SkipRecord List<ProtectorVariable> elements) throws ServiceException {
+    public List<ProtectorVariableKey> batchInsertOrUpdate(@SkipRecord List<ProtectorVariable> elements)
+            throws ServiceException {
         return crudService.batchInsertOrUpdate(elements);
     }
 
@@ -238,7 +243,8 @@ public class ProtectorVariableMaintainServiceImpl implements ProtectorVariableMa
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public PagedData<ProtectorVariable> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws ServiceException {
+    public PagedData<ProtectorVariable> lookup(String preset, Object[] objs, PagingInfo pagingInfo)
+            throws ServiceException {
         return presetLookupService.lookup(preset, objs, pagingInfo);
     }
 
@@ -254,7 +260,8 @@ public class ProtectorVariableMaintainServiceImpl implements ProtectorVariableMa
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<ProtectorVariable> lookupAsList(String preset, Object[] objs, PagingInfo pagingInfo) throws ServiceException {
+    public List<ProtectorVariable> lookupAsList(String preset, Object[] objs, PagingInfo pagingInfo)
+            throws ServiceException {
         return presetLookupService.lookupAsList(preset, objs, pagingInfo);
     }